added documentation and author annotations to files. Renamed remotes appropriately
