+++ src/components/pages/feed-components/feed-box.component.tsx
@@ -1,13 +1,18 @@
 import React from 'react';
-import { makeStyles, Box, Card, Typography } from '@material-ui/core';
+import { makeStyles, Box, Card } from '@material-ui/core';
 import { useHistory } from 'react-router';
 import { Question } from '../../../models/question';
-import * as fallbackRemote from '../../../remotes/fallback.remote';
+import * as answerRemote from '../../../remotes/answer.remote';
+import * as questionRemote from '../../../remotes/question.remote';
 import { IState } from '../../../reducers';
 import { connect } from 'react-redux';
 import { clickQuestion } from '../../../actions/question.actions';
 import { convertFromRaw, EditorState, Editor } from 'draft-js';
 
+/**
+ * @file Contains and manages questions and answer mapped into boxes within the feed container
+ * @author Keith Salzman 
+ */
 
 const drawerWidth = 100;
 const useStyles = makeStyles({
@@ -38,7 +43,7 @@ export const FeedBoxComponent: React.FC<FeedBoxComponentProps> = (props) => {
     const history = useHistory();
 
     const handleRedirectQ = async () => {
-        const retrievedAnswers = await fallbackRemote.getAnswersByQuestionId(props.question.id, 10, 0);
+        const retrievedAnswers = await answerRemote.getAnswersByQuestionId(props.question.id, 10, 0);
         props.clickQuestion(props.question);
         localStorage.setItem("questionId", JSON.stringify(props.question.id));
         localStorage.setItem("question", JSON.stringify(props.question));
@@ -47,8 +52,8 @@ export const FeedBoxComponent: React.FC<FeedBoxComponentProps> = (props) => {
     }
 
     const handleRedirectA = async () => {
-        const retrievedQuestion = await fallbackRemote.getQuestionByQuestionId(props.question.questionId);
-        const retrievedAnswers = await fallbackRemote.getAnswersByQuestionId(props.question.questionId, 10, 0);
+        const retrievedQuestion = await questionRemote.getQuestionByQuestionId(props.question.questionId);
+        const retrievedAnswers = await answerRemote.getAnswersByQuestionId(props.question.questionId, 10, 0);
         localStorage.setItem("questionId", JSON.stringify(retrievedQuestion.id));
         localStorage.setItem("question", JSON.stringify(retrievedQuestion));
         localStorage.setItem("answers", JSON.stringify(retrievedAnswers));
@@ -57,29 +62,25 @@ export const FeedBoxComponent: React.FC<FeedBoxComponentProps> = (props) => {
     }
 
     const questionContent = EditorState.createWithContent(convertFromRaw(JSON.parse(props.question.content)));
-    const onChange = () => {};
+    const onChange = () => { };
 
     //!First box here contains answers not questions, so does its handler deal with answer not questions
-    return (
-        <Box display="flex" justifyContent="center" >
-            <Card className={classes.boxInternal}>
-                {props.question.questionId ?
-                    <Box display="flex" justifyContent="center" onClick={() => handleRedirectA()}  >
-                        <Box paddingLeft={2} paddingRight={2} >
-                            <div className={classes.divInternal}><Editor editorState={questionContent} readOnly={true} onChange={onChange} /></div>
-                            <h3>{props.question.userId}</h3>
-                            <p>{props.question.creationDate}</p>
-                        </Box>
-                    </Box>
-                    :
-                    <Box>
-                        {!props.question.acceptedId && props.view === "confirm" ?
-                            <Box display="flex" justifyContent="center" onClick={() => handleRedirectQ()} >
-                                <Box paddingLeft={2} paddingRight={2}>
-                                    <Typography variant="h5">Waiting for users to accept answers ...</Typography>
-                                </Box>
+    if (!props.question.acceptedId && props.view === "confirm") {
+        return <div></div>
+    } else {
+        return (
+            <Box display="flex" justifyContent="center" >
+                <Card className={classes.boxInternal}>
+                    {props.question.questionId ?
+                        <Box display="flex" justifyContent="center" onClick={() => handleRedirectA()}  >
+                            <Box paddingLeft={2} paddingRight={2} >
+                                <div className={classes.divInternal}><Editor editorState={questionContent} readOnly={true} onChange={onChange} /></div>
+                                <h3>{props.question.userId}</h3>
+                                <p>{props.question.creationDate}</p>
                             </Box>
-                            :
+                        </Box>
+                        :
+                        <Box>
                             <Box display="flex" justifyContent="center" onClick={() => handleRedirectQ()} >
                                 <Box paddingLeft={2} paddingRight={2}>
                                     <h2>{props.question.title}</h2>
@@ -87,11 +88,12 @@ export const FeedBoxComponent: React.FC<FeedBoxComponentProps> = (props) => {
                                     <h3>{props.question.userId}</h3>
                                     <p>{props.question.creationDate}</p>
                                 </Box>
-                            </Box>}
-                    </Box>}
-            </Card>
-        </Box>
-    )
+                            </Box>
+                        </Box>}
+                </Card>
+            </Box>
+        )
+    }
 }
 
 const mapStateToProps = (state: IState) => {
